versioning-goat/requirements.txt
Flask==0.9
python-magic==0.4.3
requests==1.2.0
